Change macOS handling of undefined symbols
Newer version of XCode deprecate the ``-undefined dynamic_lookup`` flag
that pybind11 has used to avoid linking to a specific Python shared
library. Every use of it now lead to the following noisy warning
message.
```
ld: warning: -undefined dynamic_lookup may not work with chained fixups
```
More details on this are available here:
https://mjtsai.com/blog/2021/06/30/faster-app-launching-in-ios-15-and-monterey/

While this seems somewhat specific to Swift, it is problematic if a core
linker feature that pybind11 depends on is declared unreliable.

I have been able to replace these flags with
```
-undefined suppress -flat_namespace
```

I am not sure if this suits everyone. The purpose of the PR is to raise
the issue and starts a discussion on what to do.